+++ cmd/gouno/web.go
@@ -76,7 +76,7 @@ func startWebServer(cmd *cobra.Command, args []string) {
 	}
 	defer func() { _ = redis.Close() }()
 
-	auditAuditor := auditService.NewAuditor(ctx, db, logger)
+	auditAuditor := auditService.NewAuditor(ctx, db, &globalConfig.TaskPipelineConfig, logger)
 	go listenAuditErrors(ctx, auditAuditor, logger)
 
 	modules, err := initModules(ctx, db, redis, logger, globalConfig, auditAuditor)