+++ internal/audit/service/audit_test.go
@@ -28,7 +28,7 @@ func TestAudit(t *testing.T) {
 	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
 	defer cancel()
 
-	auditor := service.NewAuditor(ctx, db, nil)
+	auditor := service.NewAuditor(ctx, db, nil, nil)
 
 	testAccountID := uuid.New().String()